rm all client-api types in type-definitions
