+++ tee-worker/ts-tests/integration-tests/common/transactions.ts
@@ -6,12 +6,20 @@ import { expect } from 'chai';
 import colors from 'colors';
 import type { HexString } from '@polkadot/util/types';
 import type { Codec } from '@polkadot/types/types';
-import type { IntegrationTestContext, TransactionSubmit } from './type-definitions';
+import type { IntegrationTestContext } from './common-definitions';
 import type { ApiTypes, SubmittableExtrinsic } from '@polkadot/api/types';
-import { RequestEvent } from './type-definitions';
 
 import { u8aToHex } from '@polkadot/util';
 import { FrameSystemEventRecord } from 'parachain-api';
+
+enum RequestEvent {
+    LinkIdentityRequested = 'LinkIdentityRequested',
+    DeactivateIdentityRequested = 'DeactivateIdentityRequested',
+    ActivateIdentityRequested = 'ActivateIdentityRequested',
+    VCRequested = 'VCRequested',
+    ItemCompleted = 'ItemCompleted',
+    BatchCompleted = 'BatchCompleted',
+}
 
 // transaction utils
 export async function sendTxUntilInBlock(tx: SubmittableExtrinsic<ApiTypes>, signer: KeyringPair) {
@@ -29,7 +37,10 @@ export async function sendTxUntilInBlock(tx: SubmittableExtrinsic<ApiTypes>, sig
 
 export async function sendTxUntilInBlockList(
     api: ApiPromise,
-    txs: TransactionSubmit[],
+    txs: {
+          tx: SubmittableExtrinsic<ApiTypes>;
+    nonce: number;
+    }[],
     signer: KeyringPair | KeyringPair[]
 ) {
     const signers = Array.isArray(signer) ? signer : [signer];
@@ -159,14 +170,14 @@ export async function listenEvent(
 export async function sendTxsWithUtility(
     context: IntegrationTestContext,
     signer: KeyringPair,
-    txs: TransactionSubmit[],
+    txs: SubmittableExtrinsic<ApiTypes>[],
     pallet: string,
     events: string[],
     listenTimeoutInBlockNumber?: number
 ): Promise<Event[]> {
     // ensure the tx is in block
     const isInBlockPromise = new Promise((resolve) => {
-        context.api.tx.utility.batchAll(txs.map(({ tx }) => tx)).signAndSend(signer, async (result) => {
+        context.api.tx.utility.batchAll(txs.map((tx) => tx)).signAndSend(signer, async (result) => {
             if (result.status.isInBlock) {
                 console.log(`Transaction included at blockHash ${result.status.asInBlock}`);
                 resolve(result.status);
@@ -193,7 +204,10 @@ export async function sendTxsWithUtility(
 
 export async function multiAccountTxSender(
     context: IntegrationTestContext,
-    txs: TransactionSubmit[],
+    txs: {
+          tx: SubmittableExtrinsic<ApiTypes>;
+    nonce: number;
+    }[],
     signers: KeyringPair | KeyringPair[],
     pallet: string,
     events: string[],
