+++ tee-worker/ts-tests/integration-tests/common/utils/crypto.ts
@@ -1,7 +1,7 @@
 import type { HexString } from '@polkadot/util/types';
-import { hexToU8a, u8aToHex, stringToU8a } from '@polkadot/util';
+import { hexToU8a, stringToU8a } from '@polkadot/util';
 import { KeyObject } from 'crypto';
-import { AesOutput } from '../type-definitions';
+import { AesOutput } from 'parachain-api';
 import crypto from 'crypto';
 import { KeyringPair } from '@polkadot/keyring/types';
 import { ethers } from 'ethers';
@@ -30,7 +30,7 @@ export function encryptWithAes(key: HexString, nonce: Uint8Array, cleartext: Buf
 }
 
 export function decryptWithAes(key: HexString, aesOutput: AesOutput, type: 'hex' | 'utf-8'): HexString {
-    if (aesOutput.ciphertext && aesOutput.nonce) {
+    // if (aesOutput.ciphertext && aesOutput.nonce) {
         const secretKey = crypto.createSecretKey(hexToU8a(key));
         const tagSize = 16;
         const ciphertext = aesOutput.ciphertext ? aesOutput.ciphertext : hexToU8a('0x');
@@ -52,9 +52,9 @@ export function decryptWithAes(key: HexString, aesOutput: AesOutput, type: 'hex'
         const part2 = decipher.final(type);
 
         return `0x${part1 + part2}`;
-    } else {
-        return u8aToHex(aesOutput as Uint8Array);
-    }
+    // } else {
+        // return u8aToHex(aesOutput as Uint8Array);
+    // }
 }
 
 export interface Signer {
