+++ tee-worker/ts-tests/integration-tests/common/utils/identity-helper.ts
@@ -1,10 +1,10 @@
 import { u8aToHex } from '@polkadot/util';
 import { blake2AsHex } from '@polkadot/util-crypto';
-import type { IdentityGenericEvent, IntegrationTestContext } from '../type-definitions';
-import { AesOutput } from '../type-definitions';
+import type { IntegrationTestContext } from '../common-definitions';
+import { AesOutput } from 'parachain-api';
 import { decryptWithAes, encryptWithTeeShieldingKey, Signer } from './crypto';
 import { ethers } from 'ethers';
-import type { TypeRegistry } from '@polkadot/types';
+import type { TypeRegistry, Vec } from '@polkadot/types';
 import type { LitentryPrimitivesIdentity, PalletIdentityManagementTeeIdentityContext } from 'sidechain-api';
 import type { LitentryValidationData, Web3Network } from 'parachain-api';
 import type { ApiTypes, SubmittableExtrinsic } from '@polkadot/api/types';
@@ -153,31 +153,6 @@ export function parseIdentity(
         decryptedIdentity
     ) as unknown as LitentryPrimitivesIdentity;
     return identity;
-}
-
-export function createIdentityEvent(
-    sidechainRegistry: TypeRegistry,
-    who: HexString,
-    identityString?: HexString,
-    idGraphString?: HexString
-): IdentityGenericEvent {
-    const identity: LitentryPrimitivesIdentity =
-        identityString! &&
-        (sidechainRegistry.createType(
-            'LitentryPrimitivesIdentity',
-            identityString
-        ) as unknown as LitentryPrimitivesIdentity);
-    const idGraph: [LitentryPrimitivesIdentity, PalletIdentityManagementTeeIdentityContext][] =
-        idGraphString! &&
-        (sidechainRegistry.createType(
-            'Vec<(LitentryPrimitivesIdentity, PalletIdentityManagementTeeIdentityContext)>',
-            idGraphString
-        ) as unknown as [LitentryPrimitivesIdentity, PalletIdentityManagementTeeIdentityContext][]);
-    return <IdentityGenericEvent>{
-        who,
-        identity,
-        idGraph,
-    };
 }
 
 export async function buildValidations(
