+++ tee-worker/ts-tests/integration-tests/common/utils/vc-helper.ts
@@ -26,3 +26,86 @@ export async function handleVcEvents(
     }
     return [...results];
 }
+
+export const jsonSchema = {
+    type: 'object',
+    properties: {
+        id: {
+            type: 'string',
+        },
+        type: {
+            type: 'array',
+        },
+        issuer: {
+            type: 'object',
+            properties: {
+                id: {
+                    type: 'string',
+                },
+                name: {
+                    type: 'string',
+                },
+                shard: {
+                    type: 'string',
+                },
+            },
+        },
+        issuanceDate: {
+            type: 'string',
+        },
+        credentialSubject: {
+            type: 'object',
+            properties: {
+                id: {
+                    type: 'string',
+                },
+                description: {
+                    type: 'string',
+                },
+                type: {
+                    type: 'string',
+                },
+                tag: {
+                    type: 'array',
+                },
+                assertions: {
+                    type: 'array',
+                    items: {
+                        type: 'object',
+                    },
+                },
+                values: {
+                    type: 'array',
+                    items: {
+                        type: 'boolean',
+                    },
+                },
+                endpoint: {
+                    type: 'string',
+                },
+            },
+            required: ['id', 'description', 'type', 'assertions', 'values', 'endpoint'],
+        },
+        proof: {
+            type: 'object',
+            properties: {
+                created: {
+                    type: 'string',
+                },
+                type: {
+                    enum: ['Ed25519Signature2020'],
+                },
+                proofPurpose: {
+                    enum: ['assertionMethod'],
+                },
+                proofValue: {
+                    type: 'string',
+                },
+                verificationMethod: {
+                    type: 'string',
+                },
+            },
+        },
+    },
+    required: ['id', 'type', 'credentialSubject', 'issuer', 'issuanceDate', 'proof'],
+};
