+++ tee-worker/ts-tests/integration-tests/di_evm_identity.test.ts
@@ -23,7 +23,7 @@ import {
     sendRequestFromGetter,
     sendRequestFromTrustedCall,
 } from './common/di-utils'; // @fixme move to a better place
-import type { IntegrationTestContext } from './common/type-definitions';
+import type { IntegrationTestContext } from './common/common-definitions';
 import { aesKey } from './common/call';
 import { LitentryValidationData, Web3Network } from 'parachain-api';
 import { LitentryPrimitivesIdentity } from 'sidechain-api';
